pyproject.toml
[project]
name = "submodule-update-request"
description = "Open pull request to update submodule"
version = "0.1.0"
dependencies = [
  "click >= 8.0.4, < 9.0.0",
  "github3.py >= 3.2.0, < 4.0.0",
]

[project.optional-dependencies]
dev = [
  "black >= 22.1.0, < 23.0.0",
]

[build-system]
requires = ["flit_core >= 3.2, < 4"]
build-backend = "flit_core.buildapi"

[options]
python_requires = ">=3.9"

[tool.flit.module]
name = "sur"

[tool.black]
line-length = 79
target-version = ["py39", "py310"]
